feat(admin): reuse account menu
Share the signed-in avatar menu between user and admin headers, preserving theme and sign-out actions while removing the standalone theme toggle.
