+++ ui/src/App.tsx
@@ -744,7 +744,7 @@ function App() {
         onSignOut={signOut}
       />
       <SidebarInset className="min-h-0 overflow-hidden">
-        <SiteHeader />
+        <SiteHeader authSession={authSession} onSignOut={signOut} />
         <main className="flex min-h-0 flex-1 flex-col gap-4 overflow-y-auto p-4 lg:gap-6 lg:p-6">
           {section !== "accounts" ? (
             <div className="grid gap-4 md:grid-cols-2 xl:grid-cols-4">
