+++ ui/src/components/user-dashboard.tsx
@@ -7,25 +7,20 @@ import {
   Github,
   KeyRound,
   Loader2,
-  LogOut,
-  Monitor,
-  Moon,
   Play,
   Pencil,
   RefreshCw,
-  Settings,
   ShieldCheck,
   SquareTerminal,
-  Sun,
   Workflow,
   X,
 } from "lucide-react"
-import { useTheme } from "next-themes"
 import { type CSSProperties, type FormEvent, type MouseEvent, type ReactNode, useEffect, useMemo, useRef, useState } from "react"
 
 import type { AuthSession, GitHubAppConfig, RunnerJobGroup, RunnerState, UserPreferences } from "@/admin-types"
 import { logNames } from "@/admin-types"
 import { formatRunnerDuration, formatTime } from "@/admin-format"
+import { AccountMenu } from "@/components/account-menu"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
@@ -38,16 +33,6 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog"
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
@@ -211,7 +196,7 @@ export function UserDashboard({
           </a>
         </nav>
         <div className="ml-auto flex items-center gap-2">
-          <UserMenu authSession={authSession} onSignOut={onSignOut} />
+          <AccountMenu authSession={authSession} onSignOut={onSignOut} />
         </div>
       </header>
 
@@ -1223,75 +1208,6 @@ function PullRequestsPage({
   )
 }
 
-function UserMenu({ authSession, onSignOut }: { authSession: AuthSession; onSignOut: () => void }) {
-  const { setTheme, theme } = useTheme()
-  const avatarURL = userAvatarURL(authSession)
-  const login = authSession.login || "github"
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button type="button" variant="ghost" size="icon" className="rounded-full" aria-label="Account menu">
-          {avatarURL ? (
-            <img
-              src={avatarURL}
-              alt=""
-              className="h-8 w-8 rounded-full border bg-muted object-cover"
-              referrerPolicy="no-referrer"
-            />
-          ) : (
-            <span className="flex h-8 w-8 items-center justify-center rounded-full border bg-muted text-xs font-semibold">
-              {userInitials(login)}
-            </span>
-          )}
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-56">
-        <DropdownMenuLabel className="truncate">{login}</DropdownMenuLabel>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem asChild>
-          <a href="/account/repositories">
-            <Settings className="h-4 w-4" />
-            Settings
-          </a>
-        </DropdownMenuItem>
-        <DropdownMenuSeparator />
-        <DropdownMenuLabel className="text-xs font-medium text-muted-foreground">Theme</DropdownMenuLabel>
-        <DropdownMenuRadioGroup value={theme || "system"} onValueChange={setTheme}>
-          <DropdownMenuRadioItem value="light">
-            <Sun className="h-4 w-4" />
-            Light
-          </DropdownMenuRadioItem>
-          <DropdownMenuRadioItem value="dark">
-            <Moon className="h-4 w-4" />
-            Dark
-          </DropdownMenuRadioItem>
-          <DropdownMenuRadioItem value="system">
-            <Monitor className="h-4 w-4" />
-            System
-          </DropdownMenuRadioItem>
-        </DropdownMenuRadioGroup>
-        <DropdownMenuSeparator />
-        {authSession.role === "admin" ? (
-          <>
-            <DropdownMenuItem asChild>
-              <a href="/admin/">
-                <ShieldCheck className="h-4 w-4" />
-                Admin
-              </a>
-            </DropdownMenuItem>
-            <DropdownMenuSeparator />
-          </>
-        ) : null}
-        <DropdownMenuItem onClick={onSignOut}>
-          <LogOut className="h-4 w-4" />
-          Sign out
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  )
-}
-
 function JobField({ label, value, onOpen }: { label: string; value: ReactNode; onOpen?: () => void }) {
   return (
     <div className="grid grid-cols-[88px_minmax(0,1fr)] items-baseline gap-2">
@@ -2256,21 +2172,4 @@ function accountAvatarURL(installation: NonNullable<GitHubAppConfig["installatio
   if (installation.account_avatar) return installation.account_avatar
   if (!installation.account_login) return ""
   return `https://github.com/${encodeURIComponent(installation.account_login)}.png?size=96`
-}
-
-function userAvatarURL(authSession: AuthSession) {
-  if (authSession.avatar_url) return authSession.avatar_url
-  if (!authSession.login) return ""
-  return `https://github.com/${encodeURIComponent(authSession.login)}.png?size=96`
-}
-
-function userInitials(login: string) {
-  return (
-    login
-      .split(/[-_\s]+/)
-      .filter(Boolean)
-      .slice(0, 2)
-      .map((part) => part[0]?.toUpperCase())
-      .join("") || "GH"
-  )
 }
